pyproject.toml
[tool.poetry]
name = "mini"
version = "0.1.0"
description = ""
authors = ["Aditya Pratama Nusantara <adityaadit677@gmail.com>"]
readme = "README.md"
packages = [{include = "mini"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.19"
langchain-groq = "^0.2.4"
langchain-community = "^0.3.18"
streamlit = "^1.42.2"
langchain-google-genai = "^2.0.10"
google-genai = "^1.2.0"
arxiv = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.setuptools.packages.find]
where = ["."]
include = ["mini"]
namespaces = false
